% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classification.R
\name{calculate_probability}
\alias{calculate_probability}
\title{Calculate the genotox probability and classification.}
\usage{
calculate_probability(
  probe_values,
  genotox_scores,
  non_genotox_scores,
  probe_score_sds
)
}
\arguments{
\item{probe_values}{Vector of Log10 fold-change values for each probe for the
test chemical.}

\item{probe_score_sds}{Vector of training standard deviation values for each
probe provided by the classifier data frame.}

\item{probe_scores}{Vector of genotox negative training scores for each probe.}
}
\value{
Named list with the calculated probabilities and classification:
\itemize{
\item prob_genotoxic
\item prob_non_genotoxic
\item classification
}
}
\description{
Calculate the genotox probability and classification.
}
