% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classification.R
\name{check_classifier_version}
\alias{check_classifier_version}
\title{Verify the provided classifier version is valid.}
\usage{
check_classifier_version(classifier_version = c(1, 2))
}
\arguments{
\item{classifier_version}{Version of the trained DDI classifier to use. Possible values are 1 or 2.}
}
\description{
Verify the provided classifier version is valid.
}
