% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classification.R
\name{classify}
\alias{classify}
\title{Classify a set of test materials as genotoxic, non-genotoxic, or unclassified.}
\usage{
classify(input_data, classifier_version, include_training = FALSE)
}
\arguments{
\item{input_data}{A two column data frame that has gene symbols in column 1
and log10 fold-change data in column 2. Column 1 should be named "Probe" and
the name of column 2 will be taken as the test material identifier for labeling
purposes.}

\item{classifier_version}{Version of the trained DDI classifier to use. Possible values are 1 or 2.}

\item{include_training}{If TRUE, include the training chemicals in the classification.}
}
\value{
A list with the following members:
\itemize{
\item fc_matrix - A matrix of just the log10 fold change data for each chemical. Rows are genes and columns are chemicals.
\item classification_df - A data frame with the classification results for each test material.
Columns are:
* chem_id - The name of the test material
* prob_genotoxic - The probability of the test material being genotoxic
* prob_non_genotoxic - The probability of the test material being non-genotoxic
* classification - The classification of the test material (Genotoxic, Non-Genotoxic, or Unclassified)
\item gene_dist - The gene distance matrix
\item gene_clust_dendo - The dendogram of the gene distance matrix
}
}
\description{
Classify a set of test materials as being genotoxic, non-genotoxic, or unclassified.
Optionally include the training chemical set in the classification.
}
