% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch.R
\name{classify_batch_file_against_training}
\alias{classify_batch_file_against_training}
\title{Classify a batch of chemical results from a file against the training data.}
\usage{
classify_batch_file_against_training(
  batch_filepath,
  out_dir = NULL,
  data_start_column = 2,
  classifier_version = 2,
  plot_title = NULL,
  save_outputs = TRUE
)
}
\arguments{
\item{batch_filepath}{Path to a file containing a batch of input data to classify as tab delimited
text.}

\item{out_dir}{The output directory to place the resulting figures and tables.
Defaults to the current working directory}

\item{classifier_version}{Version of the trained DDI classifier to use. Possible values are 1 or 2.}

\item{plot_title}{The title to use in the classification figures. Defaults to NULL.}

\item{save_outputs}{If TRUE, save the resulting figures and tables to the output directory.}
}
\description{
Classify the provided test materials from a batch file and compare
against the training data.
}
