% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classification.R
\name{classify_general}
\alias{classify_general}
\title{Classify a test material as genotoxic, non-genotoxic, or unclassified.}
\usage{
classify_general(
  input_data,
  classifier_version,
  out_dir = NULL,
  outfile_prefix = NULL,
  plot_title,
  save_outputs = TRUE
)
}
\arguments{
\item{input_data}{A two column data frame that has gene symbols in column 1
and log10 fold-change data in column 2. Column 1 should be named "Probe" and
the name of column 2 will be taken as the test material identifier for labeling
purposes.}

\item{classifier_version}{Version of the trained DDI classifier to use. Possible values are 1 or 2.}

\item{out_dir}{The output directory to place the resulting figures and tables.
Defaults to the current working directory}

\item{outfile_prefix}{A prefix to use in the file names of any saved output from
this function. Defaults to the test material name found in the input_data argument.}
}
\description{
Classify a test material as being genotoxic, non-genotoxic, or unclassified
by comparing against a known training set of genotoxic and non-genotoxic
chemicals.
}
