% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classification.R
\name{create_fc_matrix}
\alias{create_fc_matrix}
\title{Create a matrix of just log10 fold change data from the provided data frame.}
\usage{
create_fc_matrix(valid_df)
}
\arguments{
\item{valid_df}{A list as returned by \link{validate_merged_input_df} describing
the validated input data}
}
\description{
Create a matrix of just log10 fold change data from the provided data frame.
}
