% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmap.R
\name{default_class_colors}
\alias{default_class_colors}
\title{Get a matrix that maps DDI classifications to colors for plotting purposes.}
\usage{
default_class_colors()
}
\description{
Builds a 2 column matrix in the following format:
\enumerate{
\item classification: The classification of the material (Genotoxic, Non-Genotoxic, Unclassified)
\item color: The color to use for the classification
}
}
