% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logging.R
\name{get_logfile_path}
\alias{get_logfile_path}
\title{Get package logfile path.}
\usage{
get_logfile_path()
}
\description{
Get the value of the LOG_FILEPATH variable stored in the package environment
if available. Will return NULL if the variable has not yet been set.
}
