% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tgx_ddi.package.R
\name{get_pkg_filepath}
\alias{get_pkg_filepath}
\title{Helper function to load a file defined by this package in the \emph{inst} directory.}
\usage{
get_pkg_filepath(...)
}
\description{
Helper function to load a file defined by this package in the \emph{inst} directory.
}
