% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tgx_ddi.package.R
\name{is.error}
\alias{is.error}
\title{Check if an object is an error}
\usage{
is.error(x)
}
\arguments{
\item{x}{An object to check}
}
\description{
Check to see if the provided object is an S3 class that inherits from "try-error"
}
