% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classification.R
\name{load_classifier}
\alias{load_classifier}
\title{Load the trained classifier version into a dataframe.}
\usage{
load_classifier(classifier_version)
}
\arguments{
\item{classifier_version}{Version of the trained DDI classifier to use. Possible values are 1 or 2.}
}
\description{
Load the trained classifier version into a dataframe.
}
