% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classification.R
\name{load_heatmap}
\alias{load_heatmap}
\title{Load the trained heatmap data into a dataframe.}
\usage{
load_heatmap(classifier_version)
}
\arguments{
\item{classifier_version}{Version of the trained DDI classifier to use. Possible values are 1 or 2.}
}
\description{
Load the trained heatmap data into a dataframe.
}
