% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classification.R
\name{load_merged_classifier}
\alias{load_merged_classifier}
\title{Load merged training data.}
\usage{
load_merged_classifier(classifier_version)
}
\description{
Load the specified trained classifier joined to the trained heatmap data by
gene symbol.
}
