% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classification.R
\name{load_training_fc}
\alias{load_training_fc}
\title{Load the log10 fold-change data for the training chemicals.}
\usage{
load_training_fc(classifier_version)
}
\arguments{
\item{classifier_version}{Version of the trained DDI classifier to use. Possible values are 1 or 2.}
}
\description{
Load the log10 fold-change data for the training chemicals.
}
