% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logging.R
\name{log_msg}
\alias{log_msg}
\title{Log a message}
\usage{
log_msg(msg, file)
}
\description{
Helper method that will log a message to the logfile (if configured) and send
echo the message to the console using the \link[base:message]{base::message} function
}
