% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmap.R
\name{plot_heatmap_against_training}
\alias{plot_heatmap_against_training}
\title{Build heatmap comparing a single test material against the training materials.}
\usage{
plot_heatmap_against_training(
  fc_matrix,
  classification_df,
  gene_clust_dendo,
  plot_title,
  timestamp
)
}
\description{
This function builds a heatmap comparing a single test material against the training materials.
The test material should be the last column of the fc_matrix.
}
