% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch.R
\name{read_batch_df}
\alias{read_batch_df}
\title{Read a batch dataframe}
\usage{
read_batch_df(batch_df, data_start_column = 2)
}
\arguments{
\item{batch_df}{Dataframe containing batch results. The first column should
be the gene symbol followed zero or more metadata columns followed by data
columns where each column represents log2 fold change data for a chemical
instance. Each of these data columns should have a column header that describes
the chemical of the format \emph{ChemicalName_Concentration_ConcentrationUnits}.}

\item{data_start_column}{Index of the column where the log2 fold change data
starts. Defaults to 2.
The incoming dataframe may have metadata columns between the gene symbol column
and the fold change columns. Build a sequence of columns of just the FC columns.}
}
\description{
Read a batch file containing results for multiple chemical instances returning
a list of dataframes: 1 dataframe per chemical instance. The log2 fold change
data will also be converted to log10 fold change for ease of subsequent analysis.
}
