% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logging.R
\name{set_logfile_path}
\alias{set_logfile_path}
\title{Set package logfile path.}
\usage{
set_logfile_path(logfile_path = NULL)
}
\description{
Set the value of the LOG_FILEPATH variable stored in the package environment.
If this value is set, all logged messages will be sent to the file specified
as well as the console.
}
