% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch.R
\name{sort_batch_df}
\alias{sort_batch_df}
\title{Sort a batch dataframe by chemical name then concentration.}
\usage{
sort_batch_df(valid_batch_df)
}
\arguments{
\item{valid_batch_df}{Validated batch dataframe as produced by \link{validate_batch_df}.}
}
\description{
Sort a batch dataframe by chemical name then concentration.
}
