% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch.R
\name{split_validated_batch_df}
\alias{split_validated_batch_df}
\title{Split a batch dataframe into a list of dataframes.}
\usage{
split_validated_batch_df(valid_batch_df)
}
\arguments{
\item{valid_batch_df}{Validated batch dataframe as produced by \link{validate_batch_df}.}
}
\value{
A list of dataframes, one for each chemical instance.
}
\description{
Read a batch data frame containing results for multiple chemical instances returning
a list of dataframes: 1 dataframe per chemical instance. The log2 fold change
data will also be converted to log10 fold change for ease of subsequent analysis.
}
