% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tgx_ddi.package.R
\docType{package}
\name{tgxddi-package}
\alias{tgxddi}
\alias{tgxddi-package}
\title{tgxddi: TGx-DDI}
\description{
The TGx-DDI biomarker was developed as a toxicogenomics signature to identify chemicals that can cause DNA damage in human cells in culture.
}
\keyword{internal}
