% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tgx_ddi.package.R
\docType{data}
\name{this_pkg}
\alias{this_pkg}
\title{Define a package wide environment for storing package variables}
\format{
An object of class \code{environment} of length 0.
}
\usage{
this_pkg
}
\description{
Define a package wide environment for storing package variables
}
\keyword{datasets}
