% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch.R
\name{validate_batch_df}
\alias{validate_batch_df}
\title{Validate a batch dataframe is in the expected format.}
\usage{
validate_batch_df(batch_df, data_start_column = 2)
}
\arguments{
\item{batch_df}{Dataframe containing batch results.}

\item{data_start_column}{Index of the column where the log2 fold change data
starts. Defaults to 2.}
}
\value{
A validated dataframe with the following form:
\enumerate{
\item The gene symbol column is converted to uppercase and named "Probe".
\item The remaining columns with the test chemical log2 fold change data converted
to log10 fold change. The columns are sorted by chemical name then concentration.
The incoming dataframe may have metadata columns between the gene symbol column
and the fold change columns. Drop these columns.
}
}
\description{
Validate the incoming batch data frame by:
\enumerate{
\item There are at least 2 columns in the dataframe.
\item The first column is the gene symbol.
\item Starting at the provided \emph{data_start_column} (default=2), assume
the remaining columns are log2 fold change data for each test chemical
instance. These columns should have a column header that describe that
uniquely labels the chemical instance of the format
\emph{ChemicalName_Concentration_ConcentrationUnits}.
}
}
