% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch.R
\name{validate_chemical_id}
\alias{validate_chemical_id}
\title{Validate a chemical identifier is in the expected format.}
\usage{
validate_chemical_id(chemical_id)
}
\arguments{
\item{chemical_id}{The chemical identifier to validate.}
}
\value{
A list containing the name, concentration, and units of the chemical.
}
\description{
Verify that the chemical identifier is in the expected format:
\emph{ChemicalName_Concentration_Units}.
}
