% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classification.R
\name{validate_merged_input_df}
\alias{validate_merged_input_df}
\title{Validate the input data frame for classification.}
\usage{
validate_merged_input_df(input_df)
}
\arguments{
\item{input_df}{A data frame that should have the following columns:
\itemize{
\item Probe
\item Genotoxic.score
\item Non.Genotoxic.score
\item my.train.sd
\item One or more columns with log10 fold-change data for each chemical instance
we'd like to predict probabilites and classification on.
}}
}
\value{
A list with the following members:
\itemize{
\item df - The unchanged validated data frame
\item probe_colname - The name of the column containing the Probe information.
\item genotoxic_score_colname - The name of the column containing positive genotox score weights.
\item non_genotoxic_score_colname - The name of the column containing negative genotox score weights.
\item sd_colname - The name of the column containing standard deviation info for each probe.
\item data_colnames - The names of the columns containing the log10 fold change data.
}
}
\description{
Validate a data frame that should be a merged combination of the result returned
from \link{load_merged_classifier} with the log10 fold change data from a test
chemical. Verifies that the data frame is in the correct shape.
}
