% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch.R
\name{classify_batch_file}
\alias{classify_batch_file}
\title{Classify a batch of test chemical results from a file.}
\usage{
classify_batch_file(
  batch_filepath,
  data_start_column = 2,
  classifier_version = 2,
  save_outputs = TRUE,
  plot_title = NULL,
  out_dir = NULL,
  outfile_prefix = NULL
)
}
\arguments{
\item{batch_filepath}{Path to a file containing a batch of input data to classify as tab delimited
text.}

\item{classifier_version}{Version of the trained DDI classifier to use. Possible values are 1 or 2.}

\item{save_outputs}{If TRUE, save the resulting figures and tables to the output directory.}

\item{plot_title}{The title to use in the classification figures. Defaults to NULL.}

\item{out_dir}{The output directory to place the resulting figures and tables.
Defaults to the current working directory}

\item{outfile_prefix}{A prefix to use in the file names of any saved output from
this function. Defaults to "tgx_ddi_results".}
}
\description{
Classif the provided test materials from a batch file but do not
compare against the training data.
}
