% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch.R
\name{read_batch_file}
\alias{read_batch_file}
\title{Read a batch TSV file.}
\usage{
read_batch_file(batch_filepath)
}
\arguments{
\item{batch_filepath}{Path to the tab separated input file.}
}
\description{
Read a batch file into a dataframe.
}
